Revert "Refactor tests"
This reverts commit 6941c7755d9915cb3ecc054995c67c2b50598adb.
