--- src/test/kotlin/com/zopa/ktor/opentracing/utils/testUtils.kt
@@ -1,6 +1,0 @@
-package com.zopa.ktor.opentracing.utils
-
-import com.zopa.ktor.opentracing.ThreadContextElementScopeManager
-import io.opentracing.mock.MockTracer
-
-val mockTracer = MockTracer(ThreadContextElementScopeManager())
